Improve lost DAT handling
The retransmit timeout used the last time a packet is received
to determine when the timeout expired. This has the problem
that when dealing with a client that implements timeouts (which
are shorter than the server's), the client resends the last
ACK incase it was lost.

To avoid the sorceror's apprentice problem, duplicate ACKs are
ignored and do not resend the DAT.

Together, this results in the problem where the client continues
to resend the ACK, and the server never resends the data.

To solve this, we add tracking when a packet is dropped, and only
uptime the timer when a non-dropped packet is received. This allows
the duplicate ACKs to be ignored, but not reset the last_packet
time stat, allowing the server to eventually timeout on its own
and resend the DAT packet.